requirements.txt
python-dotenv
fastapi
uvicorn[standard]
sqlalchemy
sqlmodel
alembic
itsdangerous

PyJWT

passlib
bcrypt==4.0.1

python-multipart
psycopg2
pandas

google-auth-oauthlib
google-auth-httplib2
sib_api_v3_sdk
redis
colorlog
asyncpg
pydantic-settings
pydantic[email]
httpx
resend

botocore>=1.31.0
aioboto3==12.3.0
boto3>=1.28.0,<2.0.0
openpyxl
moto
xlrd
aiofiles
aiohttp
celery[redis]>=5.3.0,<6

sympy
openpyxl

gspread
google-auth
google-api-python-client
pandas
esdk-obs-python

tenacity

flower


zipstream-ng
zipstream-new


nest-asyncio
clear
nest_asyncio
azure-storage-blob==12.27.1
azure-identity
